pyproject.toml
[project]
name = "quantization-adalora-speech-recognition"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = "==3.13.*"
dependencies = [
    "accelerate>=1.12.0",
    "bitsandbytes>=0.49.2",
    "datasets>=4.8.3",
    "huggingface-hub>=1.4.1",
    "librosa>=0.10.0",
    "numpy>=1.24.0",
    "pandas>=2.0.0",
    "peft>=0.18.1",
    "python-dotenv>=1.2.1",
    "rich>=14.3.2",
    "soundfile>=0.13.1",
    "torch==2.8.0",
    "torchao==0.9.0",
    "torchaudio>=2.8.0",
    "torchcodec>=0.10.0",
    "transformers>=4.57.6",
    "typer>=0.24.0",
]

[tool.uv.sources]
torch = [
    { index = "pytorch-cu126", marker = "sys_platform != 'darwin'" },
    { index = "pytorch-cpu", marker = "sys_platform == 'darwin'" },
]
torchao = [
    { index = "pytorch-cu126", marker = "sys_platform != 'darwin'" },
    { index = "pytorch-cpu", marker = "sys_platform == 'darwin'" },
]

[[tool.uv.index]]
name = "pytorch-cu126"
url = "https://download.pytorch.org/whl/cu126"
explicit = true

[[tool.uv.index]]
name = "pytorch-cpu"
url = "https://download.pytorch.org/whl/cpu"
explicit = true
